Sparse.Key

Documentation

newtype Key

key i j interleaves the bits of the keys i and j.

Keys are then just values sorted in "Morton Order".

Constructors

Key

Fields

runKey :: Word64

key :: Word32 -> Word32 -> Key

This should be order preserving

shuffled :: Iso' (Word32, Word32) Key

unshuffled :: Iso' Key (Word32, Word32)

_ii :: Lens' Key Word64

Lenses for working with masked off versions of the i or j column. The result is spread with interleaved zeros in the odd bits.

_jj :: Lens' Key Word64

Lenses for working with masked off versions of the i or j column. The result is spread with interleaved zeros in the odd bits.