Sparse.Matrix

Description

Sparse Matrices in Morton order

Sparse Matrices

newtype Mat v a

Constructors

Mat

Fields

runMat :: Vector Vector v (Key, a)

Keys

data Key

key i j interleaves the bits of the keys i and j.

Keys are then just values sorted in "Morton Order".

key :: Word32 -> Word32 -> Key

Construct a key from a pair of indices.

 key i j ^. _1 = i
 key i j ^. _2 = j

shuffled :: Iso' (Word32, Word32) Key

This isomorphism lets you build a key from a pair of indices.

 key i j ≡ (i,j)^.shuffled

 shuffled . unshuffled = id
 unshuffled . shuffled = id

unshuffled :: Iso' Key (Word32, Word32)

This isomorphism lets you build a pair of indices from a key.

Construction

fromList :: Vector v a => [(Key, a)] -> Mat v a

Build a sparse matrix.

singleton :: Vector v a => Key -> a -> Mat v a

singleton makes a matrix with a singleton value at a given location

ident :: (Vector v a, Num a) => Word32 -> Mat v a

ident n makes an n x n identity matrix

empty :: Vector v a => Mat v a

The empty matrix

Consumption

count :: Mat v a -> Int

Count the number of non-zero entries in the matrix

Distinguishable Zero

class Eq0 a where

Methods

isZero :: a -> Bool

Customization

addWith :: Vector v a => (a -> a -> Maybe a) -> Mat v a -> Mat v a -> Mat v a

Merge two matrices where the indices coincide into a new matrix. This provides for generalized addition. Return Nothing for zero.

multiplyWith :: Vector v a => (a -> a -> a) -> (a -> a -> Maybe a) -> Mat v a -> Mat v a -> Mat v a

Multiply two matrices using the specified multiplication and addition operation.

We can work with the Boolean semiring as a Mat Data.Vector.Unboxed.Vector () using:

 booleanOr = addWith (const . Just)
 booleanAnd = multiplyWith const (const . Just)

nonZero :: Eq0 c => (a -> b -> c) -> a -> b -> Maybe c

Remove results that are equal to zero from a simpler function.

When used with addWith or multiplyWith's additive argumnt this can help retain the sparsity of the matrix.

Lenses

_Mat :: Iso (Mat u a) (Mat v b) (Vector Vector u (Key, a)) (Vector Vector v (Key, b))

This isomorphism lets you access the internal structure of a matrix

keys :: Lens' (Mat v a) (Vector Key)

Access the keys of a matrix

values :: Lens (Mat u a) (Mat v b) (u a) (v b)

Access the keys of a matrix