Sparse.Matrix.Fusion

Description

Matrix stream fusion internals

Documentation

mergeStreamsWith :: (Monad m, Ord i) => (a -> a -> Maybe a) -> Stream m (i, a) -> Stream m (i, a) -> Stream m (i, a)

This is the internal stream fusion combinator used to merge streams for addition.

timesSingleton :: Monad m => (a -> b -> c) -> Stream m (Key, a) -> (Key, b) -> Stream m (Key, c)

This is the internal stream fusion combinator used to multiply on the right by a singleton key

singletonTimes :: Monad m => (a -> b -> c) -> (Key, a) -> Stream m (Key, b) -> Stream m (Key, c)

This is the internal stream fusion combinator used to multiply on the left by a singleton key