Sparse.Matrix.Internal.Fusion

Description

Matrix stream fusion internals

Documentation

mergeStreamsWith :: Monad m => (a -> a -> a) -> Stream m (Key, a) -> Stream m (Key, a) -> Stream m (Key, a)

This is the internal stream fusion combinator used to merge streams for addition.

mergeStreamsWith0 :: Monad m => (a -> a -> Maybe a) -> Stream m (Key, a) -> Stream m (Key, a) -> Stream m (Key, a)

This is the internal stream fusion combinator used to merge streams for addition.

This form permits cancellative addition.