Sparse.Matrix.Key

Description

Keys in Morton order

This module provides combinators for shuffling together the bits of two key components to get a key that is based on their interleaved bits.

See http://en.wikipedia.org/wiki/Z-order_curve for more information about Morton order.

Documentation

newtype Key

key i j interleaves the bits of the keys i and j.

Keys are then just values sorted in "Morton Order".

Constructors

Key

Fields

runKey :: Word64

key :: Word32 -> Word32 -> Key

Construct a key from a pair of indices.

 key i j ^. _1 = i
 key i j ^. _2 = j

shuffled :: Iso' (Word32, Word32) Key

This isomorphism lets you build a key from a pair of indices.

 key i j ≡ (i,j)^.shuffled

 shuffled . unshuffled = id
 unshuffled . shuffled = id

unshuffled :: Iso' Key (Word32, Word32)

This isomorphism lets you build a pair of indices from a key.