Sparse.Matrix

Description

Sparse Matrices in Morton order

Sparse Matrices

data Mat v a

Constructors

Mat !Int !(Vector Word) !(Vector Word) !(v a)

Keys

data Key

Key i j logically orders the keys as if the bits of the keys i and j were interleaved. This is equivalent to storing the keys in "Morton Order".

>>> Key 100 200 ^. _1
100

>>> Key 100 200 ^. _2
200

Constructors

Key !Word !Word

Construction

fromList :: Vector v a => [(Key, a)] -> Mat v a

Build a sparse matrix.

singleton :: Vector v a => Key -> a -> Mat v a

singleton makes a matrix with a singleton value at a given location

transpose :: Vector v a => Mat v a -> Mat v a

Transpose a matrix

ident :: (Vector v a, Num a) => Int -> Mat v a

ident n makes an n x n identity matrix

empty :: Vector v a => Mat v a

The empty matrix

Consumption

size :: Mat v a -> Int

Count the number of non-zero entries in the matrix

Distinguishable Zero

class Num a => Eq0 a where

Methods

isZero :: a -> Bool

Return whether or not the element is 0.

It may be okay to never return True, but you won't be able to thin spurious zeroes introduced into your matrix.

addMats :: Vector v a => Mat v a -> Mat v a -> Mat v a

Add two matrices. By default this assumes isZero can possibly return True after an addition. For some ring-like structures, this doesn't hold. There you can use:

 addMats = addWith (+)

addHeap :: Maybe (Heap a) -> Stream (Key, a)

Convert from a Heap to a Stream.

If addition of non-zero valus in your ring-like structure cannot yield zero, then you can use

 addHeap = Heap.streamHeapWith (+)

instead of the default definition

 addHeap = Heap.streamHeapWith0 $ nonZero (+)

Customization

addWith :: Vector v a => (a -> a -> a) -> Mat v a -> Mat v a -> Mat v a

Merge two matrices where the indices coincide into a new matrix. This provides for generalized addition. Return Nothing for zero.

multiplyWith :: Vector v a => (a -> a -> a) -> (Maybe (Heap a) -> Stream (Key, a)) -> Mat v a -> Mat v a -> Mat v a

Multiply two matrices using the specified multiplication and addition operation.

nonZero :: Eq0 c => (a -> b -> c) -> a -> b -> Maybe c

Remove results that are equal to zero from a simpler function.

When used with addWith or multiplyWith's additive argumnt this can help retain the sparsity of the matrix.

Lenses

_Mat :: Iso (Mat u a) (Mat v b) (Vector Vector u (Key, a)) (Vector Vector v (Key, b))

bundle up the matrix in a form suitable for vector-algorithms

keys :: Lens' (Mat v a) (Vector Key)

Access the keys of a matrix

values :: Lens (Mat u a) (Mat v b) (u a) (v b)

Access the keys of a matrix